remove test suffix from base test class
